+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   the root action is added to the `Beacon`.
 - Fix `enterAction` for a ended/left `RootAction` to return a `NullAction` with the `RootAction` as parent instead of
   a `NullAction` without parent.
+- Add IOpenKit::createSession overload without IP address parameter.  
+  The IP address is determined in this case on the server side.
 
 ## 1.1.5 [Release date: 2019-08-08]
 [GitHub Releases](https://github.com/Dynatrace/openkit-native/releases/tag/v1.1.5)